pyarrow_stubs_ext/pyarrow_stubs_ext/pa_array_types.py
from typing import (
    TypeVar,
)
import pyarrow as pa

PaArray = pa.Array | pa.ChunkedArray
StructArrayOrChunk = pa.StructArray | pa.ChunkedArray

PaDataT = TypeVar('PaDataT', bound=PaArray | pa.Table)
PaArrayT = TypeVar('PaArrayT', bound=PaArray)
