pyproject.toml
[tool.poetry]
name = "pyarrow-stubs"
version = "0.1"
description = "Type annotations for pyarrow"
authors = ["Dale Jung <dale@dalejung.com>"]
packages = [
    { "include" = "pyarrow-stubs"}
]

[tool.poetry.dependencies]
pyarrow_stubs_ext = { path = "./pyarrow_stubs_ext", develop = true}

[tool.pyright]
typeCheckingMode = "strict"
include = ["pyarrow-stubs"]

# disable subset of strict
reportIncompleteStub = false
reportInconsistentConstructor = false
reportMissingParameterType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportUnusedExpression = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUntypedBaseClass = false
reportUnusedVariable = false
reportPrivateUsage = false
reportInvalidStubStatement = false
# enable optional checks
reportMissingModuleSource = true
